refactor: reorganize repository imports for deployed agents and sessions
