+++ apps/electron/src/main/infrastructure/database/core/repository-factory.ts
@@ -1,9 +1,9 @@
 import { SqliteManager } from "./sqlite-manager";
 import { AgentRepository } from "../repositories/agent/agent-repository";
-import { DeployedAgentRepository } from "../repositories/deployed-agent/deployed-agent-repository";
+import { DeployedAgentRepository } from "../repositories/agent/deployed-agent-repository";
 import { LogRepository } from "../repositories/log/log-repository";
 import { ServerRepository } from "../repositories/server/server-repository";
-import { SessionRepository } from "../repositories/session/session-repository";
+import { SessionRepository } from "../repositories/agent/session-repository";
 import { SettingsRepository } from "../../../modules/settings/settings.repository";
 import { TokenRepository } from "../repositories/token/token-repository";
 import { WorkspaceRepository } from "../../../modules/workspace/workspace.repository";
