+++ apps/electron/src/main/infrastructure/database/repositories/agent/deployed-agent-repository.ts
