+++ apps/electron/src/main/infrastructure/database/repositories/agent/session-repository.ts
